Поправка пиксельперфекта и автоматизации
