+++ gulpfile.js
@@ -5,6 +5,7 @@ var less = require("gulp-less");
 var plumber = require("gulp-plumber");
 var postcss = require("gulp-postcss");
 var autoprefixer = require("autoprefixer");
+var server = require("browser-sync").create();
 var minify = require("gulp-csso");
 var rename = require("gulp-rename");
 var imagemin = require("gulp-imagemin");
@@ -12,10 +13,9 @@ var webp = require("gulp-webp");
 var svgstore = require("gulp-svgstore");
 var posthtml = require("gulp-posthtml");
 var include = require("posthtml-include");
-var server = require("browser-sync").create();
 var run = require("run-sequence");
 var del = require("del");
-var jsMinify = require("gulp-uglify");
+var uglify = require("gulp-uglify");
 
 gulp.task("style", function() {
   gulp.src("less/style.less")
@@ -25,18 +25,12 @@ gulp.task("style", function() {
       autoprefixer()
     ]))
     .pipe(gulp.dest("build/css"))
+    .pipe(server.stream())
     .pipe(minify())
     .pipe(rename("style.min.css"))
-    .pipe(gulp.dest("build/css"))
-    .pipe(server.stream());
+    .pipe(gulp.dest("build/css"));
 });
 
-gulp.task("js", function() {
-  gulp.src("js/script.js")
-    .pipe(jsMinify())
-    .pipe(rename("script.min.js"))
-    .pipe(gulp.dest("build/js"))
-})
 
 gulp.task("serve", function() {
   server.init({
@@ -52,31 +46,25 @@ gulp.task("serve", function() {
 });
 
 gulp.task("images", function() {
-  return gulp.src("build/img/**/*{png,jpg,svg}")
+  return gulp.src("img/*.{png,jpg,svg}")
     .pipe(imagemin([
-      imagemin.optipng({
-        optimizationLevel: 3
-      }),
-      imagemin.jpegtran({
-        progressive: true
-      }),
+      imagemin.optipng({optimizationLevel: 3}),
+      imagemin.jpegtran({progressive: true}),
       imagemin.svgo()
     ]))
-    .pipe(gulp.dest("build/img"))
+    .pipe(gulp.dest("img"));
 });
 
 gulp.task("webp", function() {
-  return gulp.src("build/img/**/*.{png,jpg}")
-    .pipe(webp({
-      quality: 90
-    }))
-    .pipe(gulp.dest("build/img"));
+  return gulp.src("img/**/*.{png,ipg}")
+    .pipe(webp({quality: 90}))
+    .pipe(gulp.dest("img"));
 });
 
 gulp.task("sprite", function() {
-  return gulp.src("img/inline-*.svg")
+  return gulp.src("img/icon-*.svg")
     .pipe(svgstore({
-      inlineSvg: true
+      inLineSvg: true
     }))
     .pipe(rename("sprite.svg"))
     .pipe(gulp.dest("build/img"));
@@ -92,19 +80,36 @@ gulp.task("html", function() {
 
 gulp.task("copy", function() {
   return gulp.src([
-      "fonts/**/*.{woff,woff2}",
-      "img/**",
-      "js/**"
-    ], {
-      base: "."
-    })
+    "fonts/**/*.{woff,woff2}",
+    "img/**",
+    "js/**"
+  ], {
+    base: "."
+  })
     .pipe(gulp.dest("build"));
-});
-
-gulp.task("build", function(done) {
-  run("clean", "copy", "images", "webp", "style", "js", "sprite", "html", done);
 });
 
 gulp.task("clean", function() {
   return del("build");
+});
+
+
+gulp.task("scripts", function() {
+  return gulp.src("js/*.js")
+    .pipe(uglify())
+    .pipe(rename("script.min.js"))
+    .pipe(gulp.dest("build/js"));
+});
+
+
+gulp.task("build", function(done) {
+  run(
+    "clean",
+    "copy",
+    "style",
+    "sprite",
+    "scripts",
+    "html",
+    done
+  );
 });
