+++ index.html
@@ -5,7 +5,7 @@
   <meta charset="utf-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>HTML Academy: MISHKA</title>
-  <link href="css/style.css" rel="stylesheet">
+  <link href="css/style.min.css" rel="stylesheet">
   <script>
     document.createElement( "picture" );
   </script>
@@ -386,8 +386,7 @@
     </div>
   </section>
   <div class="goods-popup__overlay"></div>
-  <script type="text/javascript" src="js/script-menu.js"></script>
-  <script type="text/javascript" src="js/script-popup.js"></script>
+  <script type="text/javascript" src="js/script.min.js"></script>
 </body>
 
 </html>
