+++ js/script.js
@@ -1,3 +1,22 @@
+//меню
+var button = document.querySelector(".page-header__button");
+var menu = document.querySelector(".page-header__navigation");
+
+menu.classList.remove("page-header__navigation--no-js");
+
+button.addEventListener("click", function(event) {
+  event.preventDefault();
+  if (menu.classList.contains("page-header__navigation--closed")) {
+    menu.classList.add("page-header__navigation--opened");
+    menu.classList.remove("page-header__navigation--closed");
+  }
+  else {
+    menu.classList.add("page-header__navigation--closed");
+    menu.classList.remove("page-header__navigation--opened");
+  }
+});
+
+//поп-ап
 var overlay = document.querySelector(".goods-popup__overlay");
 var modal = document.querySelector(".goods-popup");
 var goodsPopularButton = document.querySelector(".goods-popular__button");
