+++ less/blocks/about-us.less
@@ -19,7 +19,7 @@
 
 .about-us__item {
   position: relative;
-  margin-top: 85px;
+  margin-top: 84px;
   color: @text-color--dark;
   font-size: 17px;
   line-height: 20px;
@@ -51,7 +51,7 @@
 
   .about-us__item /*+ для компьютера*/ {
     width: 304px;
-    margin-bottom: 40px;
+    margin-bottom: 42px;
     margin-right: 30px;
     padding-top: 50px;
     font-size: 20px;
