+++ less/blocks/offer-form.less
@@ -84,14 +84,6 @@
       padding-left: 18px;
       padding-bottom: 0;
 
-      &::-moz-placeholder {
-        color: fade(@text-color, 30%);
-      }
-
-      &::-webkit-input-placeholder {
-        color: fade(@text-color, 30%);
-      }
-
       &:hover {
         border: 2px solid @aqua;
       }
@@ -305,7 +297,7 @@ input[type="checkbox"]:checked + .offer-form__button--checkbox-text {
       justify-content: space-between;
       flex-direction: row-reverse;
       min-height: 110px;
-      margin-top: 25px;
+      margin-top: 23px;
     }
   }
 
@@ -331,6 +323,7 @@ input[type="checkbox"]:checked + .offer-form__button--checkbox-text {
     margin-left: 130px;
     padding-top: 40px;
     padding-bottom: 20px;
+    width: 430px;
 
     &--color {
       display: flex;
