+++ less/blocks/offer.less
@@ -27,7 +27,7 @@
       left: 0;
       z-index: 2;
       width: 100%;
-      height: 24px;
+      height: 32px;
       background-image: url("../img/bg-zigzag-fill.svg");
       background-repeat: repeat-x;
       background-size: 750px 50px;
