+++ less/blocks/page-footer.less
@@ -46,7 +46,7 @@
   &--html {
     fill: @aqua;
     top: -60px;
-    left: 180px;
+    left: 190px;
     top: -12px;
   }
 
@@ -82,7 +82,7 @@
   line-height: 18px;
   font-weight: bold;
   color: @text-color--dark;
-  margin-top: 20px;
+  margin-top: 30px;
   padding-left: 40px;
 
   span {
@@ -153,7 +153,7 @@
   }
 
   .page-footer__copyright-link {
-    margin-top: 55px;
+    margin-top: 60px;
     padding-left: 0;
   }
 }
