+++ less/blocks/page-header.less
@@ -46,7 +46,7 @@
 
   .page-header__logo  {
     margin-right: 50px;
-    margin-left: 30px;
+    margin-left: 39px;
   }
 
   .page-header__navigation--closed {
