+++ less/blocks/products.less
@@ -14,7 +14,7 @@
 
 .products__item {
   border: 1px solid @grey-93;
-  margin-bottom: 28px;
+  margin-bottom: 30px;
   min-height: 374px;
 
   &:hover {
@@ -181,7 +181,7 @@
   .products__item {
     display: inline-block;
     width: 317px;
-    height: 662px;
+    min-height: 662px;
     margin-bottom: 60px;
 
     &:last-child {
@@ -195,7 +195,7 @@
 
   .products__image-link {
     width: 316px;
-    height: 462px;
+    min-height: 462px;
   }
 
   .products__description {
