+++ less/blocks/reviews.less
@@ -30,6 +30,7 @@
   font-size: 25px;
   font-weight: bold;
   line-height: 30px;
+  margin-bottom: 23px;
 }
 
 .reviews__wrapper {
@@ -38,7 +39,7 @@
 
 .reviews__description {
   margin: 0;
-  padding-bottom: 25px;
+  padding-bottom: 27px;
   font-size: 15px;
   line-height: 30px;
 }
@@ -149,7 +150,7 @@
   }
 
   .reviews__description {
-    padding-top: 12px;
+    padding-top: 24px;
     padding-bottom: 40px;
     font-size: 17px;
     line-height: 30px;
@@ -225,7 +226,6 @@
   }
 
   .reviews__description {
-    padding-top: 5px;
     padding-bottom: 40px;
     font-size: 17px;
     line-height: 30px;
