+++ package.json
@@ -7,7 +7,7 @@
     "@htmlacademy/editorconfig-cli": "1.x",
     "autoprefixer": "7.1.x",
     "browser-sync": "2.18.x",
-    "del": "^3.0.0",
+    "del": "3.0.0",
     "gulp": "3.9.x",
     "gulp-csso": "^3.0.0",
     "gulp-imagemin": "^3.4.0",
@@ -15,20 +15,25 @@
     "gulp-plumber": "1.1.x",
     "gulp-postcss": "7.0.x",
     "gulp-posthtml": "^3.0.0",
-    "gulp-rename": "^1.2.2",
+    "gulp-rename": "1.2.2",
     "gulp-svgstore": "^6.1.0",
     "gulp-uglify": "^3.0.0",
+    "gulp-uglyfly": "^1.4.2",
     "gulp-webp": "^2.3.0",
-    "posthtml-include": "^1.1.0",
+    "posthtml-include": "1.1.0",
     "run-sequence": "^2.2.0",
     "stylelint": "8.1.x",
     "stylelint-config-htmlacademy": "0.1.x"
   },
   "scripts": {
     "editorconfig": "editorconfig-cli",
-    "stylelint": "stylelint 'less/**/*.less' —syntax less",
+    "stylelint": "stylelint 'less/**/*.less' --syntax less",
     "test": "npm run editorconfig && npm run stylelint",
     "build": "gulp build",
+    "images": "gulp images",
+    "webp": "gulp webp",
+    "sprite": "gulp sprite",
+    "scripts": "gulp uglify",
     "start": "npm run build && gulp serve"
   },
   "browserslist": [
@@ -46,5 +51,10 @@
   ],
   "engines": {
     "node": "8.2"
+  },
+  "dependencies": {
+    "del": "^3.0.0",
+    "gulp-rename": "^1.2.2",
+    "posthtml-include": "^1.1.0"
   }
 }
